% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_baseline.R
\name{analyze_baseline}
\alias{analyze_baseline}
\title{Analyze data from a baseline assessment of hemodynamics and/or sympathetic
nerve activity.}
\usage{
analyze_baseline(data)
}
\arguments{
\item{data}{a data frame output from \link{run_baseline} to be summarised.}
}
\value{
a data frame containing summary statistics for all variables.
}
\description{
The required object for this function is the output of \link{run_baseline}
}
