% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_chamber.R
\name{analyze_chamber}
\alias{analyze_chamber}
\title{Analyze data from a lower body negative pressure trial}
\usage{
analyze_chamber(data, drop_stimulus = T)
}
\arguments{
\item{data}{a data frame output from \link{run_chamber} to be summarised.}

\item{drop_stimulus}{a logical, if \code{TRUE} the initial respond to the
chamber pressure will be ignored in the summary statistics. If \code{FALSE}
the entire trial will be used in the calculation of summary statistics.
Default is \code{TRUE}.}
}
\value{
a data frame containing summary statistics for all variables.
}
\description{
The required object for this function is the output of \link{run_chamber}.
}
