% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\docType{data}
\name{baseline}
\alias{baseline}
\title{An example dataset of an uninjured animal baseline physiological parameters.}
\format{
a data frame object with dimensions 309150 rows
}
\usage{
data(baseline)
}
\description{
The data is provided as a data frame.
}
\keyword{datasets}
