% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_artifacts.R
\name{clean_artifacts}
\alias{clean_artifacts}
\title{Function to clean artifacts and NAs in a file loaded by \link{load_labchart}}
\usage{
clean_artifacts(data, n_sd = 5)
}
\arguments{
\item{data}{a data frame to be cleaned.}

\item{n_sd}{the threshold to define an artifact. Default is 5.
to}
}
\value{
a cleaned data frame.
}
\description{
Specify a threshold for artifacts and check for NA/NaNs in the data
}
