% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Function for outlier detection in physiological data}
\usage{
detect_outliers(data, rate = 1000, bin_length = 10, variable = "SBP")
}
\arguments{
\item{data}{a data frame containing only numeric columns to be cleaned.}

\item{rate}{a numeric, the sampling rate of the data. Note if the data was
downsampled out of \code{LabChart} the corrected sampling rate should be
specified in frames per second. Default is 1000.}

\item{bin_length}{the size of the binning window. Default is 10 seconds.}

\item{variable}{the variable to use for outlier removal. Default is SBP.}
}
\value{
a cleaned data frame with outliers removed.
}
\description{
Detect and remove outliers in physiological data.
}
