% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_labchart.R
\name{load_labchart}
\alias{load_labchart}
\title{Load txt files exported from LabChart}
\usage{
load_labchart(file, col_names = NULL)
}
\arguments{
\item{file}{an exported file of type \code{.txt} from \code{LabChart}
to be processed.}

\item{col_names}{a vector specifying the column names of the file. These
correspond to the channels defined in \code{LabChart}.}
}
\value{
a tibble with column names set to \code{col_names}.
}
\description{
Convert an exported LabChart text file to a user-friendly structure that is
compatible with downstream functions from Aima.
}
