% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chamber_trace.R
\name{plot_chamber_trace}
\alias{plot_chamber_trace}
\title{Plot a line graph of a lower-body negative pressure trial.}
\usage{
plot_chamber_trace(
  data,
  features = NULL,
  outcome = "delta",
  palette = "Greys",
  combine_plots = T
)
}
\arguments{
\item{data}{a data frame of lower-body negative pressure data obtained from
\link{run_chamber}.}

\item{features}{the features to plot. A separate plot will be provided for
each feature. The feature must be present as a column name of \code{data}.}

\item{outcome}{the type of data to show. Please select from \code{raw}
or \code{delta}. Defaults to \code{delta}.}

\item{palette}{the color palette used to visualize the trace for each variable
type. Can be provided either a string specifying a ColorBrewer (see
\code{\link{scale_color_brewer}}) or viridis (see
\code{\link{scale_color_viridis}}) color palette, or a vector of
colors that will be provided to \code{\link{scale_color_manual}}}

\item{combine_plots}{optionally, combine the plots into a single object.
Please note that this will prevent users from adding ggplot layers.
Defaults to \code{TRUE}.}
}
\value{
a \code{ggplot} object.
}
\description{
Visualize the change in a hemodynamic or sympathetic nervous system variable
that is specified in the output of \link{run_chamber}.
}
\details{
The following color maps can be specified using the \code{palette} argument:
\enumerate{
  \item viridis color scales: magma, inferno, plasma, viridis, or cividis
    (the default)
  \item sequential ColorBrewer scales: BuGn, BuPu, GnBu, Greens, Greys,
    Oranges, OrRd, PuBu, PuBuGn, PuRd, Purples, RdPu, Reds, YlGn, YlGnBu,
    YlOrBr, or YlOrRd
  \item diverging ColorBrewer scales: BrBG, PiYG, PRGn, PuOr, RdBu, RdGy,
    RdYlBu, RdYlGn, or Spectral
  \item a vector of colors that will be passed into
    \code{scale_color_manual}
}
}
