% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chamber.R
\name{run_chamber}
\alias{run_chamber}
\title{Analyze data from a lower body negative pressure trial.}
\usage{
run_chamber(
  data,
  rate = 1000,
  normalize_time = T,
  baseline_length = 60,
  trial_length = 120,
  n_steps = 2,
  step_length = 60,
  chamber_col = "Chamber",
  initial_chamber_pressure = 10,
  chamber_step = 5
)
}
\arguments{
\item{data}{a data frame output from \link{load_labchart} to be analyzed.}

\item{rate}{a numeric, the sampling rate of the data. Note if the data was
downsampled out of \code{LabChart} the corrected sampling rate should be
specified in frames per second. Default is 1000.}

\item{normalize_time}{a logical, if \code{TRUE} data will be interpolated
to one datapoint per second using an internal function \link{match_interp}.
If \code{FALSE} the existing sampling rate will be retained. Default is
\code{TRUE}.}

\item{baseline_length}{a numeric, the length of time in seconds the user
would like to use for the calculation of baseline statistics prior to
the onset of the chamber. Default is 60 seconds.}

\item{trial_length}{a numeric, the length of time in seconds the user
would like to specify that the trial proceeded for following the onset
of the chamber. Default is 120 seconds.}

\item{n_steps}{a numeric, the number of chamber steps used in this
lower-body negative pressure trial. Default is 2 steps.}

\item{step_length}{a numeric, the length of time in seconds that each
step was sustained for prior to proceeding to the subsequent step or the
trial finishing. Default is 60 seconds.}

\item{chamber_col}{a character, the name of the column containing the analog
data of the chamber pressure. This will be used to identify the start
and stop points of the chamber assessment. Default is 'Chamber'.}

\item{initial_chamber_pressure}{a numeric, the intial pressure specified
for the lower-body negative pressure chamber. This value will be used
to annotate the meta data. Default is 10 mmHg.}

\item{chamber_step}{a numeric, the size of pressure change between each
step of the chamber during the trial. Default is 5 mmHg.}
}
\value{
a data frame containing the lower-body negative pressure data with
  all the required meta data annotated.
}
\description{
The required object for this function is the output of \link{load_labchart}.
}
